+++ crates/fluvio-sc-schema/src/lib.rs
@@ -7,7 +7,7 @@ pub mod smartmodule;
 pub mod partition;
 pub mod versions;
 pub mod objects;
-pub mod table;
+pub mod tableformat;
 pub mod smartstream;
 
 mod apis;